add confident score

+++ Models.py
@@ -22,8 +22,7 @@ elif torch.backends.mps.is_available():
     device = torch.device('mps')
 else:
     device = torch.device('cpu')
-# device = torch.device('cpu')
-dtype = torch.float32
+dtype = torch.float16
 
 
 class DNN(nn.Module):
